@@ -1,6 +1,6 @@
 {
   "name": "iotize-toolbox",
-  "version": "0.0.4",
+  "version": "0.0.5",
   "author": "IoTize Solutions",
   "homepage": "https://www.iotize.com",
   "scripts": {
@@ -18,21 +18,22 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/common": "^7.2.2",
-    "@angular/core": "^7.2.2",
-    "@angular/forms": "^7.2.2",
+    "@angular/common": "~8.1.2",
+    "@angular/core": "~8.1.2",
+    "@angular/forms": "~8.1.2",
     "@angular/http": "^7.2.2",
-    "@angular/platform-browser": "^7.2.2",
-    "@angular/platform-browser-dynamic": "^7.2.2",
-    "@angular/router": "^7.2.2",
+    "@angular/platform-browser": "~8.1.2",
+    "@angular/platform-browser-dynamic": "~8.1.2",
+    "@angular/router": "~8.1.2",
     "@ionic-native/core": "^5.0.0",
     "@ionic-native/dialogs": "^5.15.1",
     "@ionic-native/keyboard": "^5.5.1",
     "@ionic-native/nfc": "^5.6.1",
+    "@ionic-native/secure-storage": "^5.16.0",
     "@ionic-native/splash-screen": "^5.0.0",
     "@ionic-native/status-bar": "^5.0.0",
-    "@ionic/angular": "^4.1.0",
-    "@iotize/cordova-plugin-iotize-ble": "^1.1.0-alpha.33",
+    "@ionic/angular": "^4.7.1",
+    "@iotize/cordova-plugin-iotize-ble": "^1.1.0-alpha.34",
     "@iotize/device-client.js": "0.0.1-alpha.105",
     "@iotize/device-com-nfc.cordova": "^1.0.0-alpha.12",
     "@iotize/ng-com-services": "0.0.1",
@@ -42,8 +43,9 @@
     "cordova-plugin-app-version": "0.1.9",
     "cordova-plugin-device": "^2.0.2",
     "cordova-plugin-dialogs": "^2.0.2",
-    "cordova-plugin-ionic-keyboard": "^2.1.3",
-    "cordova-plugin-ionic-webview": "^4.0.1",
+    "cordova-plugin-ionic-keyboard": "^2.2.0",
+    "cordova-plugin-ionic-webview": "^4.1.3",
+    "cordova-plugin-secure-storage": "^3.0.2",
     "cordova-plugin-splashscreen": "^5.0.2",
     "cordova-plugin-statusbar": "^2.4.2",
     "cordova-plugin-whitelist": "^1.3.3",
@@ -53,20 +55,20 @@
     "zone.js": "~0.8.29"
   },
   "devDependencies": {
-    "@angular-devkit/architect": "~0.13.8",
-    "@angular-devkit/build-angular": "^0.13.9",
-    "@angular-devkit/core": "~7.3.8",
-    "@angular-devkit/schematics": "~7.3.8",
-    "@angular/cli": "^7.3.9",
-    "@angular/compiler": "~7.2.2",
-    "@angular/compiler-cli": "~7.2.2",
-    "@angular/language-service": "~7.2.2",
-    "@ionic/angular-toolkit": "~1.5.1",
+    "@angular-devkit/architect": "~0.801.2",
+    "@angular-devkit/build-angular": "~0.801.2",
+    "@angular-devkit/core": "~8.1.2",
+    "@angular-devkit/schematics": "~8.1.2",
+    "@angular/cli": "~8.1.2",
+    "@angular/compiler": "~8.1.2",
+    "@angular/compiler-cli": "~8.1.2",
+    "@angular/language-service": "~8.1.2",
+    "@ionic/angular-toolkit": "^2.1.1",
     "@iotize/device-com-mock.js": "0.0.1-alpha.16",
     "@types/jasmine": "^3.4.4",
     "@types/jasminewd2": "^2.0.8",
     "@types/node": "~10.14.2",
-    "codelyzer": "~4.5.0",
+    "codelyzer": "~5.0.0",
     "jasmine-core": "^3.5.0",
     "jasmine-spec-reporter": "~4.2.1",
     "karma": "^4.4.1",
@@ -77,7 +79,7 @@
     "protractor": "~5.4.0",
     "ts-node": "~8.1.0",
     "tslint": "~5.16.0",
-    "typescript": "~3.1.6"
+    "typescript": "~3.4.3"
   },
   "description": "An Ionic project",
   "cordova": {
@@ -87,14 +89,15 @@
       "cordova-plugin-statusbar": {},
       "cordova-plugin-device": {},
       "cordova-plugin-splashscreen": {},
+      "cordova-plugin-app-version": {},
+      "cordova-plugin-dialogs": {},
+      "cordova-plugin-secure-storage": {},
       "cordova-plugin-ionic-webview": {
         "ANDROID_SUPPORT_ANNOTATIONS_VERSION": "27.+"
       },
       "cordova-plugin-ionic-keyboard": {},
-      "cordova-plugin-app-version": {},
       "cordova-plugin-iotize-device-com-nfc": {},
-      "cordova-plugin-iotize-ble": {},
-      "cordova-plugin-dialogs": {}
+      "cordova-plugin-iotize-ble": {}
     },
     "platforms": [
       "ios",
