@@ -14,6 +14,7 @@ import { Keyboard } from '@ionic-native/keyboard/ngx';
 import { NFC } from '@ionic-native/nfc/ngx';
 import { ComponentsModule } from './components/components.module';
 import { Dialogs } from '@ionic-native/dialogs/ngx';
+import { SecureStorage } from '@ionic-native/secure-storage/ngx';
 
 @NgModule({
   declarations: [AppComponent],
@@ -31,7 +32,8 @@ import { Dialogs } from '@ionic-native/dialogs/ngx';
     { provide: RouteReuseStrategy, useClass: IonicRouteStrategy },
     Keyboard,
     NFC,
-    Dialogs
+    Dialogs,
+    SecureStorage
   ],
   bootstrap: [AppComponent]
 })
