@@ -1,14 +1,13 @@
 import { Injectable } from '@angular/core';
 import { IoTizeTap } from '@iotize/ng-com-services';
 import { ComProtocol, ConnectionState } from '@iotize/device-client.js/protocol/api';
-import { TargetProtocol } from '@iotize/device-client.js/device/model';
 import { NFCComProtocol } from '@iotize/device-com-nfc.cordova';
 import { DiscoveredDeviceType } from '@iotize/cordova-plugin-iotize-ble';
 import { Events, LoadingController, ToastController, Platform } from '@ionic/angular';
 import { NFCTag } from './nfc.service';
-import { ResultCodeTranslation } from '@iotize/device-client.js/client/api/response';
 import { ComService } from './com.service';
 import { Subscription, timer, BehaviorSubject } from 'rxjs';
+import { SecureStorage, SecureStorageObject } from '@ionic-native/secure-storage/ngx';
 
 @Injectable({
   providedIn: 'root'
@@ -24,34 +23,44 @@ export class TapService {
     return this.iotizeTap.sessionStateForceUpdate();
   }
 
-  private selectedDevice?: DiscoveredDeviceType;
+  private _selectedDevice?: DiscoveredDeviceType;
+  private _secureStore?: SecureStorageObject;
+  private _isReconnecting = false;
+  private _keepAlivePending = false;
+  private _reconnectionSubscription?: Subscription;
+
+  public didReconnect = false;
 
   constructor(public iotizeTap: IoTizeTap,
     public events: Events,
     public loading: LoadingController,
     public toast: ToastController,
     public platform: Platform,
-    public com: ComService) {
+    public com: ComService,
+    private secureStorage: SecureStorage) {
 
     this._connectionState$ = new BehaviorSubject(ConnectionState.DISCONNECTED);
     // Save the currently connected tap on NFC Pairing
-    this.events.subscribe('NFCPairing', (tag: DiscoveredDeviceType) => this.selectedDevice = tag);
+    this.events.subscribe('NFCPairing', (tag: DiscoveredDeviceType) => this._selectedDevice = tag);
     // Allow new connection if a TAP has been discovered and no tap is connected
     this.events.subscribe('tag-discovered', (tag: NFCTag) => {
-      if (!this.tap || !this.tap.isConnected() || !this.isReady) {
+      if (!this.tap || !this.tap.isConnected() || !this.isReady || this._selectedDevice.name == tag.appName) { // if this is the same tap, 
         this.NFCLoginAndBLEPairing(tag);
       }
     });
 
     this._handleAppClosing();
   }
+
   get tap() {
     return this.iotizeTap.tap;
   }
 
-  async init(comProtocol: ComProtocol, selectedDevice?: DiscoveredDeviceType, nfcPairing = false) {
+  async init(comProtocol: ComProtocol, _selectedDevice?: DiscoveredDeviceType, nfcPairing = false) {
     try {
+      this.didReconnect = false;
       await this.iotizeTap.init(comProtocol);
+      this.handleUnexpectedDisconnection();
     } catch (err) {
       //retry
       try {
@@ -80,8 +89,8 @@ export class TapService {
     //   throw new Error("Tap is not configured for serial or modbus communication");
     // }
     // await this.tap.service.target.connect();
-    if (selectedDevice) {
-      this.selectedDevice = selectedDevice
+    if (_selectedDevice) {
+      this._selectedDevice = _selectedDevice
     }
 
     if (!nfcPairing) { // if init is called while nfc pairing, launch keep alive and sync connection state after switching to BLE
@@ -91,17 +100,26 @@ export class TapService {
   }
 
   async disconnect() {
+    this.didReconnect = false;
+    if (this._reconnectionSubscription) {
+      this._reconnectionSubscription.unsubscribe();
+
+    }
     if (this._keepAliveSubscription) {
       this._keepAliveSubscription.unsubscribe(); // stops keepAlive periodic polling
+    }
+    if (this._connectionStateSubscription) {
+      this._connectionStateSubscription.unsubscribe();
     }
     if (this.tap) {
       try {
         await this.tap.disconnect();
+        this.deleteCredentials();
       } catch (err) {
         console.error(err);
       }
     }
-    this.selectedDevice = null;
+    this._selectedDevice = null;
     if (this._connectionState$.value != ConnectionState.DISCONNECTED)
       this._connectionState$.next(ConnectionState.DISCONNECTED);
   }
@@ -115,10 +133,17 @@ export class TapService {
   }
 
   login(user: string, password: string) {
-    return this.iotizeTap.login(user, password);
+    return this.iotizeTap.login(user, password).then(status => {
+      this.events.publish('logged-in');
+      if (status) {
+        this.storeCredentials(user, password);
+      }
+      return status;
+    });
   }
 
   logout() {
+    this.deleteCredentials();
     return this.iotizeTap.logout();
   }
 
@@ -235,35 +260,58 @@ export class TapService {
 
   isConnected(device: DiscoveredDeviceType) {
     let isConnected = false;
-    if (this.tap && this.tap.isConnected() && this.selectedDevice) {
-      isConnected = device.address == this.selectedDevice.address
+    if (this.tap && this.tap.isConnected() && this._selectedDevice) {
+      isConnected = device.address == this._selectedDevice.address
     }
     return isConnected;
   }
 
   private _keepAliveSubscription?: Subscription;
   keepAliveRoutine(period: number = 1000) {
+    this._keepAlivePending = false;
     if (this._keepAliveSubscription) {
       this._keepAliveSubscription.unsubscribe();
     }
     this._keepAliveSubscription = timer(0, period).subscribe(
-      () => {
-        if (this.tap.isConnected()) {
-          this.tap.service.interface.keepAlive().catch(
-            err => {
-              if (err.code && err.code == "ConnectionError")
-                this.disconnect();
-            }
-          );
+      async () => {
+        if (this.tap.isConnected() && !this._keepAlivePending && !this._isReconnecting) {
+          this._keepAlivePending = true;
+          try {
+            await this.tap.service.interface.keepAlive();
+          } catch (error) {
+            console.error('Keep Alive', error);
+          }
+          this._keepAlivePending = false;
         }
       }
     )
   }
 
+  private async _reconnect() {
+    const loader = await this.loading.create({
+      message: 'Connection lost. Reconnecting...'
+    });
+    loader.present();
+    console.warn('RECONNECTING');
+    this._isReconnecting = true;
+    try {
+      await this.tap.connect();
+      this.didReconnect = true;
+      // disable encrypted communication and reset sessionKey and IV app-side (destroyed Tap-side) 
+      await this.tap.logout();
+      await this.loginFromStorage();
+    } catch (error) {
+      console.warn('Could not reconnect to the tap');
+      console.warn(error);
+    }
+    this._isReconnecting = false;
+    loader.dismiss();
+  }
+
   private _handleAppClosing() {
     window.addEventListener('beforeunload', () => {
-      if (this.selectedDevice) {
-        console.log('disconnecting from', this.selectedDevice.name);
+      if (this._selectedDevice) {
+        console.log('disconnecting from', this._selectedDevice.name);
         this.disconnect();
       }
     });
@@ -289,9 +337,77 @@ export class TapService {
 
   }
 
+  private handleUnexpectedDisconnection() {
+    this._reconnectionSubscription = this.connectionState.subscribe(
+      state => {
+        console.debug(ConnectionState[state]);
+        if (state == ConnectionState.DISCONNECTED
+          && this._selectedDevice
+          && !this._isReconnecting
+          && this._keepAliveSubscription) {
+
+          this._reconnect();
+        }
+      }
+    )
+  }
+
   exit() {
-    this.tap.disconnect();
-    this._keepAliveSubscription.unsubscribe();
-    this._connectionStateSubscription.unsubscribe();
+    this.disconnect();
+  }
+
+  private async storeCredentials(login: string, password: string) {
+    if (!this._secureStore) {
+      console.log('can not store password: phone is not secured');
+      return;
+    }
+    try {
+      await Promise.all(
+        [
+          this._secureStore.set('login', login),
+          this._secureStore.set('password', password)
+        ]
+      );
+      console.log('credentials stored');
+    } catch (error) {
+      console.debug(error);
+    }
+  }
+
+  private async deleteCredentials() {
+    if (!this._secureStore) {
+      console.log('can not store password: phone is not secured');
+      return;
+    }
+    try {
+      await Promise.all(
+        [
+          this._secureStore.remove('login'),
+          this._secureStore.remove('password')
+        ]
+      );
+    } catch (error) {
+      console.debug(error);
+    }
+  }
+
+  private async loginFromStorage() {
+    if (!this._secureStore) {
+      console.log('can not store password: phone is not secured');
+      return false;
+    }
+    try {
+      console.log('retrieving credentials from secure storage');
+      const login = await this._secureStore.get('login');
+      const password = await this._secureStore.get('password');
+      return this.login(login, password);
+    } catch (err) {
+      console.debug(err);
+      return false;
+    }
+  }
+
+  async initSecureStorage() {
+    this._secureStore = await this.secureStorage.create('toolbox');
   }
 }
