@@ -1,12 +1,11 @@
 import { Injectable } from '@angular/core';
-import { ActivatedRouteSnapshot, RouterStateSnapshot } from '@angular/router';
-import { CanDeactivate } from '@angular/router/src/utils/preactivation';
+import { ActivatedRouteSnapshot, RouterStateSnapshot, CanDeactivate } from '@angular/router';
 import { SettingsPage } from './pages/settings/settings.page';
 
 @Injectable({
   providedIn: 'root'
 })
-export class SettingsGuard implements CanDeactivate{
+export class SettingsGuard implements CanDeactivate<SettingsPage> {
   component: SettingsPage;
   route: ActivatedRouteSnapshot;
 
