@@ -15,6 +15,6 @@ getTestBed().initTestEnvironment(
   platformBrowserDynamicTesting()
 );
 // Then we find all the tests.
-const context = require.context('./', true, /infos\.page\.spec\.ts$/);
+const context = require.context('./', true, /modbus-view\.page\.spec\.ts$/);
 // And load the modules.
 context.keys().map(context);
